refactor(nlpgo): extract httpblock template into shared engine package
Lift the Liquid-subset renderer out of httpblock into
app/engine/template so the upcoming signature-node wiring can render
`instructions` and user-prompt content with the same syntax HTTP-block
customers already rely on. httpblock keeps a thin re-export wrapper so
existing callsites (executor.go, tests) don't change.

Pure refactor — same algorithm, same warnings list, no behavior
change. Same-package httpblock tests still cover RenderTemplate.
